server/requirements.txt
flask
inference-sdk
flask-cors
flask-sqlalchemy
flask-login
flask-migrate
pymysql
python-dotenv
werkzeug
gunicorn